requirements.txt
# UI
fastapi
uvicorn
python-multipart

# Core Python utilities
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Audio processing
openai-whisper

# Image & OCR (future use)
opencv-python
pillow
pytesseract

# GenAI / NLP
transformers>=4.35.0
sentencepiece
spacy
ollama
matplotlib

# Logging & helpers
loguru